+++ database/postgres.go
@@ -2,10 +2,8 @@ package database
 
 import (
 	"database/sql"
-	"strconv"
 	"time"
 
-	"github.com/buger/jsonparser"
 	_ "github.com/lib/pq"
 	"github.com/negbie/logp"
 	"github.com/sipcapture/heplify-server/config"
@@ -14,9 +12,9 @@ import (
 )
 
 type Postgres struct {
-	db      *sql.DB
-	dbTimer time.Duration
-	bulkCnt int
+	db              *sql.DB
+	dbTimer         time.Duration
+	bulkCnt         int
 	forceHEPPayload []int
 }
 
@@ -53,7 +51,7 @@ func (p *Postgres) setup() error {
 	p.bulkCnt = config.Setting.DBBulk
 
 	/* force JSON payload to data header */
-	p.forceHEPPayload = config.Setting.ForceHEPPayload;
+	p.forceHEPPayload = config.Setting.ForceHEPPayload
 
 	if p.bulkCnt < 1 {
 		p.bulkCnt = 1
@@ -105,7 +103,7 @@ func (p *Postgres) insert(hCh chan *decoder.HEP) {
 			bpd := bytebufferpool.Get()
 
 			if pkt.ProtoType == 1 && pkt.Payload != "" && pkt.SIP != nil {
-				pHeader := makeProtoHeader(pkt, pkt.SIP.XCallID, bpp)
+				pHeader := makeProtoHeader(pkt, bpp)
 				dHeader := makeSIPDataHeader(pkt, bpd)
 				switch pkt.SIP.CseqMethod {
 				case "INVITE", "UPDATE", "BYE", "ACK", "PRACK", "REFER", "CANCEL", "INFO":
@@ -134,7 +132,7 @@ func (p *Postgres) insert(hCh chan *decoder.HEP) {
 					}
 				}
 			} else if pkt.ProtoType == 54 && pkt.Payload != "" {
-				pHeader := makeProtoHeader(pkt, pkt.CID, bpp)
+				pHeader := makeProtoHeader(pkt, bpp)
 				sid, dHeader := makeISUPDataHeader([]byte(pkt.Payload), bpd)
 
 				isupRows = append(isupRows, sid, date, pHeader, dHeader, pkt.Payload)
@@ -146,7 +144,7 @@ func (p *Postgres) insert(hCh chan *decoder.HEP) {
 				}
 
 			} else if pkt.ProtoType >= 2 && pkt.Payload != "" && pkt.CID != "" {
-				pHeader := makeProtoHeader(pkt, "", bpp)
+				pHeader := makeProtoHeader(pkt, bpp)
 				dHeader := makeRTCDataHeader(pkt, bpd)
 				switch pkt.ProtoType {
 				case 5:
@@ -182,17 +180,16 @@ func (p *Postgres) insert(hCh chan *decoder.HEP) {
 					for _, v := range p.forceHEPPayload {
 						if pkt.ProtoType == uint32(v) {
 							ForcePayload = true
-							break;
+							break
 						}
-					}					
-					
+					}
+
 					if ForcePayload {
 						reportRows = append(reportRows, pkt.CID, date, pHeader, pkt.Payload, dHeader)
 					} else {
 						reportRows = append(reportRows, pkt.CID, date, pHeader, dHeader, pkt.Payload)
 					}
-					
-					
+
 					reportCnt++
 					if reportCnt == p.bulkCnt {
 						p.bulkInsert(reportCopy, reportRows)
@@ -296,157 +293,4 @@ func (p *Postgres) bulkInsert(query string, rows []string) {
 	}
 
 	//logp.Debug("sql", "%s\n\n%v\n\n", query, rows)
-}
-
-func makeProtoHeader(h *decoder.HEP, corrID string, sb *bytebufferpool.ByteBuffer) string {
-	sb.WriteString(`{`)
-	sb.WriteString(`"protocolFamily":`)
-	sb.WriteString(strconv.FormatUint(uint64(h.Version), 10))
-	sb.WriteString(`,"protocol":`)
-	sb.WriteString(strconv.FormatUint(uint64(h.Protocol), 10))
-	sb.WriteString(`,"srcIp":"`)
-	sb.WriteString(h.SrcIP)
-	sb.WriteString(`","dstIp":"`)
-	sb.WriteString(h.DstIP)
-	sb.WriteString(`","srcPort":`)
-	sb.WriteString(strconv.FormatUint(uint64(h.SrcPort), 10))
-	sb.WriteString(`,"dstPort":`)
-	sb.WriteString(strconv.FormatUint(uint64(h.DstPort), 10))
-	sb.WriteString(`,"timeSeconds":`)
-	sb.WriteString(strconv.FormatUint(uint64(h.Tsec), 10))
-	sb.WriteString(`,"timeUseconds":`)
-	sb.WriteString(strconv.FormatUint(uint64(h.Tmsec), 10))
-	sb.WriteString(`,"payloadType":`)
-	sb.WriteString(strconv.FormatUint(uint64(h.ProtoType), 10))
-	sb.WriteString(`,"captureId":"`)
-	sb.WriteString(h.NodeName)
-	if h.NodePW != "" {
-		sb.WriteString(`","capturePass":"`)
-		sb.WriteString(h.NodePW)
-	}
-	if corrID != "" {
-		sb.WriteString(`","correlation_id":"`)
-		sb.WriteString(corrID)
-	}
-	sb.WriteString(`"}`)
-	return sb.String()
-}
-
-func makeRTCDataHeader(h *decoder.HEP, sb *bytebufferpool.ByteBuffer) string {
-	sb.WriteString(`{`)
-	sb.WriteString(`"node":"`)
-	sb.WriteString(h.NodeName)
-	sb.WriteString(`","host":"`)
-	sb.WriteString(h.HostTag)
-	sb.WriteString(`"}`)
-	return sb.String()
-}
-
-var IsupPaths = [][]string{
-	[]string{"cic"},
-	[]string{"dpc"},
-	[]string{"opc"},
-	[]string{"msg_name"},
-	[]string{"called_number", "num"},
-	[]string{"calling_number", "num"},
-}
-
-func makeISUPDataHeader(data []byte, sb *bytebufferpool.ByteBuffer) (string, string) {
-	var msg_name, called_number, calling_number, callid string
-	var cic, dpc, opc int64
-
-	jsonparser.EachKey(data, func(idx int, value []byte, vt jsonparser.ValueType, err error) {
-		switch idx {
-		case 0:
-			if cic, err = jsonparser.ParseInt(value); err != nil {
-				logp.Warn("%v", err)
-			}
-		case 1:
-			if dpc, err = jsonparser.ParseInt(value); err != nil {
-				logp.Warn("%v", err)
-			}
-		case 2:
-			if opc, err = jsonparser.ParseInt(value); err != nil {
-				logp.Warn("%v", err)
-			}
-		case 3:
-			if msg_name, err = jsonparser.ParseString(value); err != nil {
-				logp.Warn("%v", err)
-			}
-		case 4:
-			if called_number, err = jsonparser.ParseString(value); err != nil {
-				logp.Warn("%v", err)
-			}
-		case 5:
-			if calling_number, err = jsonparser.ParseString(value); err != nil {
-				logp.Warn("%v", err)
-			}
-		}
-	}, IsupPaths...)
-	scic := strconv.FormatInt(cic, 10)
-	sdpc := strconv.FormatInt(dpc, 10)
-	sopc := strconv.FormatInt(opc, 10)
-	//snprintf("%d:%d:%d", opc < dpc ? opc : dpc, dpc < opc ? opc : dpc , cic)
-
-	if opc < dpc {
-		callid = sopc + ":"
-	} else {
-		callid = sdpc + ":"
-	}
-	if dpc < opc {
-		callid += sopc + ":" + scic
-	} else {
-		callid += sdpc + ":" + scic
-	}
-
-	sb.WriteString(`{`)
-	sb.WriteString(`"cic":`)
-	sb.WriteString(scic)
-	sb.WriteString(`,"dpc":`)
-	sb.WriteString(sdpc)
-	sb.WriteString(`,"opc":`)
-	sb.WriteString(sopc)
-	sb.WriteString(`,"msg_name":"`)
-	sb.WriteString(msg_name)
-	sb.WriteString(`","called_number":"`)
-	sb.WriteString(called_number)
-	sb.WriteString(`","calling_number":"`)
-	sb.WriteString(calling_number)
-	sb.WriteString(`","callid":"`)
-	sb.WriteString(callid)
-	sb.WriteString(`"}`)
-
-	return callid, sb.String()
-}
-
-func makeSIPDataHeader(h *decoder.HEP, sb *bytebufferpool.ByteBuffer) string {
-	sb.WriteString(`{`)
-	sb.WriteString(`"ruri_domain":"`)
-	sb.WriteString(h.SIP.URIHost)
-	sb.WriteString(`","ruri_user":"`)
-	sb.WriteString(h.SIP.URIUser)
-	sb.WriteString(`","from_user":"`)
-	sb.WriteString(h.SIP.FromUser)
-	sb.WriteString(`","to_user":"`)
-	sb.WriteString(h.SIP.ToUser)
-	sb.WriteString(`","pid_user":"`)
-	sb.WriteString(h.SIP.PaiUser)
-	sb.WriteString(`","auth_user":"`)
-	sb.WriteString(h.SIP.AuthUser)
-	if len(h.SIP.CHeader) > 0 {
-		for k, v := range h.SIP.CustomHeader {
-			sb.WriteString(`","` + k + `":"`)
-			sb.WriteString(v)
-		}
-	}
-	sb.WriteString(`","callid":"`)
-	sb.WriteString(h.SIP.CallID)
-	sb.WriteString(`","method":"`)
-	sb.WriteString(h.SIP.FirstMethod)
-	if h.SIP.UserAgent != "" {
-		sb.WriteString(`","user_agent":"`)
-		sb.WriteString(h.SIP.UserAgent)
-	}
-	sb.WriteString(`"}`)
-	return sb.String()
 }
